refactor: move helper functions into utils.py in groot
